V1.0.0 Fixed parameter types

+++ Firmware/SMT_Oven/Project_Headers/pit.h
@@ -210,7 +210,7 @@ public:
     *  @param interval  Interval in timer ticks (usually bus clock period)
     *  @param tctrl     Timer Control Register value
     */
-   static void configureChannel(const uint8_t channel, int interval, uint32_t tctrl=PIT_TCTRL_DEFAULT_VALUE) {
+   static void configureChannel(const uint8_t channel, uint32_t interval, uint32_t tctrl=PIT_TCTRL_DEFAULT_VALUE) {
       pit->CHANNEL[channel].LDVAL = interval;
       pit->CHANNEL[channel].TCTRL = tctrl;
       pit->CHANNEL[channel].TFLG  = PIT_TFLG_TIF_MASK;
@@ -225,7 +225,7 @@ public:
     *  @param tctrl     Timer Control Register value
     */
    static void configureChannel(const uint8_t channel, float interval, uint32_t tctrl=PIT_TCTRL_DEFAULT_VALUE) {
-      configureChannel(channel, (int)round(interval*PitInfo::getClockFrequency()), tctrl);
+      configureChannel(channel, (uint32_t)round(interval*PitInfo::getClockFrequency()), tctrl);
    }
    /**
     * Set period in seconds
@@ -260,7 +260,7 @@ public:
       while (pit->CHANNEL[channel].TFLG == 0) {
          __NOP();
       }
-      configureChannel(channel, 0, 0);
+      configureChannel(channel, (uint32_t)0, 0);
    }
 
 };
